图片验证码生成器/requirements.txt
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tortoise-orm==0.20.1
asyncpg==0.27.0